separate validateFileCount into 2 fns
