+++ server/src/utils/convertFilesToDeleteIdsToArr.ts
@@ -1,0 +1,13 @@
+export default function convertFilesToDeleteIdsToArr(filesToDeleteIds: string | string[] | undefined) {
+  let result: string[];
+
+  if (typeof filesToDeleteIds === "string") {
+    result = [filesToDeleteIds];
+  } else if (!filesToDeleteIds) {
+    result = [];
+  } else {
+    result = JSON.parse(JSON.stringify(filesToDeleteIds));
+  }
+
+  return result;
+}
